+++ src/app/np-utility-demo/np-utility-demo.module.ts
@@ -5,7 +5,7 @@ import { NpUtilityDemoRoutingModule } from './np-utility-demo-routing.module';
 import { NpUtilityDemoComponent } from './np-utility-demo.component';
 import {
   NpUtilityModule, NpHightLightPipe, NpDatePickerModule, NpTimePickerModule, NpSwitchModule,
-  NpFileUploadModule, NpCardModule, NpInputTextModule
+  NpFileUploadModule, NpCardModule, NpInputTextModule, NpTextareaModule
 } from 'np-ui-lib';
 import { FormsModule } from '@angular/forms';
 
@@ -21,6 +21,7 @@ import { FormsModule } from '@angular/forms';
     NpInputTextModule,
     NpSwitchModule,
     NpFileUploadModule,
+    NpTextareaModule,
     NpCardModule
   ],
   providers: [NpHightLightPipe]