+++ projects/np-ui-lib/src/lib/np-checkbox/np-checkbox.component.ts
@@ -22,7 +22,7 @@ export class NpCheckboxComponent implements ControlValueAccessor {
   @Input() indeterminate: boolean;
   @Input() readOnly: boolean;
   @Input() autoFocus: boolean;
-  @Input() tabIndex = 0;
+  @Input() tabIndex: number;
   @Input() styleClass: string;
   @Input() inputId = `np-checkbox_${NpCheckboxComponent.controlCount++}`;