+++ projects/np-ui-lib/src/lib/np-color-picker/np-color-picker.component.ts
@@ -28,7 +28,7 @@ export class NpColorPickerComponent implements ControlValueAccessor, AfterViewIn
   @Input() defaultOpen: boolean;
   @Input() readOnly: boolean;
   @Input() autoFocus: boolean;
-  @Input() tabIndex = 0;
+  @Input() tabIndex: number;
   @Input() styleClass: string;
   @Input() inputId = `np-color-picker_${NpColorPickerComponent.controlCount++}`;