+++ projects/np-ui-lib/src/lib/np-file-upload/np-file-upload.component.ts
@@ -33,7 +33,7 @@ export class NpFileUploadComponent implements ControlValueAccessor, Validator {
   @Input() showFileSize = true;
   @Input() readOnly: boolean;
   @Input() autoFocus: boolean;
-  @Input() tabIndex = 0;
+  @Input() tabIndex: number;
   @Input() styleClass: string;
   @Input() inputId = `np-file-upload_${NpFileUploadComponent.controlCount++}`;