+++ projects/np-ui-lib/src/lib/np-number-box/np-number-box.component.ts
@@ -33,7 +33,7 @@ export class NpNumberBoxComponent implements ControlValueAccessor, Validator {
   @Input() placeholder = '';
   @Input() readOnly: boolean;
   @Input() autoFocus: boolean;
-  @Input() tabIndex = 0;
+  @Input() tabIndex: number;
   @Input() styleClass: string;
   @Input() inputId = `np-number-box_${NpNumberBoxComponent.controlCount++}`;