+++ projects/np-ui-lib/src/lib/np-radio-button/np-radio-button.component.ts
@@ -22,7 +22,7 @@ export class NpRadioButtonComponent implements ControlValueAccessor {
   @Input() value: any;
   @Input() readOnly: boolean;
   @Input() autoFocus: boolean;
-  @Input() tabIndex = 0;
+  @Input() tabIndex: number;
   @Input() styleClass: string;
   @Input() inputId = `np-radio-button_${NpRadioButtonComponent.controlCount++}`;