+++ projects/np-ui-lib/src/lib/np-switch/np-switch.component.ts
@@ -22,7 +22,7 @@ export class NpSwitchComponent implements ControlValueAccessor {
   @Input() falseLabelText: string;
   @Input() readOnly: boolean;
   @Input() autoFocus: boolean;
-  @Input() tabIndex = 0;
+  @Input() tabIndex: number;
   @Input() styleClass: string;
   @Input() inputId = `np-switch_${NpSwitchComponent.controlCount++}`;