+++ projects/np-ui-lib/CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## 10.6.2
 - Add resize property to textarea
+- Resolved row selection issues in data grid
 
 ## 10.6.1
 - Resolved issue in object compare utility for tags and dropdown